+++ server/routes/deletePlaylist.js
@@ -1,13 +1,11 @@
 import express from 'express';
 const router = express.Router();
 
-// 플레이리스트 삭제 API
 router.delete('/deletePlaylist', async (req, res) => {
   const { playlistId } = req.body;
   const database = req.database;
 
   try {
-    // MongoDB에서 플레이리스트 데이터 삭제
     const result = await database.collection('playListData').deleteOne({ id: playlistId });
 
     if (result.deletedCount === 1) {
