+++ server/routes/getPlaylist.js
@@ -1,13 +1,11 @@
 import express from 'express';
 const router = express.Router();
 
-// 특정 플레이리스트 아이디로 플레이리스트 데이터 조회 API
 router.get('/getPlaylist', async (req, res) => {
   const { playlistId } = req.query;
   const database = req.database;
 
   try {
-    // MongoDB에서 플레이리스트 데이터 조회
     const playlist = await database.collection('playListData').findOne({ id: playlistId });
 
     if (playlist) {
